+++ example/test_files/b.f90
@@ -1,0 +1,3 @@
+subroutine b
+   use a
+end subroutine b